fixes highlighting of search result
